requirements.txt
matplotlib>=3.3.2
Cython>=0.29.21
scikit-learn==0.23.2
opencv-python==4.5.1.48
proj==0.2.0
pillow==8.0.1
xarray>=0.16.1
netcdf4>=1.5.3
google-cloud-storage>=1.37.1
google-cloud-vision>=2.3.1
pyproj>=3.3.0
pyclipper
labelme
psutil
open-fonts==0.1
tensorflow==2.7.0
keras==2.7.0
pynimbus
geopy>=2.2.0
metpy>=1.1.0
scipy==1.6.3
shapely==1.8.0
proplot
geopandas
suncalc
requests